+++ packages/extensions/core/src/lib/pipeline/plugins/loaders/index.ts
@@ -1,0 +1,2 @@
+export * from "./openapi-loader";
+export * from "./swagger-loader";
